Integrate post history test harness
